fixed player color replication, moved global variables from shootertypes to gameinstance, fixed compilation for dedicated server and for steam

+++ Source/ShooterGame/Private/Player/ShooterCheatManager.cpp
@@ -51,7 +51,7 @@ void UShooterCheatManager::ForceMatchStart()
 	}
 }
 
-void UShooterCheatManager::ChangeTeam(int32 NewTeamNumber)
+void UShooterCheatManager::ChangeTeam(uint8 NewTeamNumber)
 {
 	AShooterPlayerController* MyPC = GetOuterAShooterPlayerController();
